refactor: move StreamElement and StreamState to room.go
