+++ pkg/internal/request/create_test.go
@@ -3,7 +3,7 @@ package request
 import (
 	"bytes"
 	"encoding/json"
-	"syncstream-server/pkg/internal/stream"
+	"syncstream-server/pkg/internal/room"
 	"testing"
 	"time"
 
@@ -29,7 +29,7 @@ func TestCreateHandler(t *testing.T) {
 			req: createReqBodyToRequest(CreateRequestBody{
 				ID:  uuid.New(),
 				URL: "https://example.com",
-				StreamState: stream.StreamState{
+				StreamState: room.StreamState{
 					CurrentTime:  314.159,
 					Paused:       true,
 					PlaybackRate: 1.0,
