+++ pkg/internal/request/join_test.go
@@ -3,7 +3,6 @@ package request
 import (
 	"log/slog"
 	"syncstream-server/pkg/internal/room"
-	"syncstream-server/pkg/internal/stream"
 	"testing"
 	"time"
 
@@ -16,7 +15,7 @@ import (
 )
 
 func TestJoinHandler(t *testing.T) {
-	ss := stream.StreamState{
+	ss := room.StreamState{
 		CurrentTime:  314.159,
 		Paused:       true,
 		PlaybackRate: 1.0,
